Renamed featureB Api and Impl packages to foxFeature with the new task
