+++ app/src/main/java/com/featuremodule/template/ui/AppNavHost.kt
@@ -8,8 +8,8 @@ import androidx.navigation.navigation
 import com.featuremodule.core.navigation.NavBarItems
 import com.featuremodule.featureAApi.FeatureADestination
 import com.featuremodule.featureAImpl.registerFeatureA
-import com.featuremodule.featureBApi.FoxFeatureDestination
-import com.featuremodule.featureBImpl.registerFoxFeature
+import com.featuremodule.foxFeatureApi.FoxFeatureDestination
+import com.featuremodule.foxFeatureImpl.registerFoxFeature
 import com.featuremodule.homeApi.HomeDestination
 import com.featuremodule.homeImpl.registerHome
