+++ feature/featureBApi/src/main/java/com/featuremodule/foxFeatureApi/FoxFeatureDestination.kt
@@ -1,4 +1,4 @@
-package com.featuremodule.featureBApi
+package com.featuremodule.foxFeatureApi
 
 object FoxFeatureDestination {
     const val ROUTE = "fox_feature"
