+++ feature/featureBImpl/src/main/java/com/featuremodule/foxFeatureImpl/FoxFeatureGraphEntry.kt
@@ -1,9 +1,9 @@
-package com.featuremodule.featureBImpl
+package com.featuremodule.foxFeatureImpl
 
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
-import com.featuremodule.featureBApi.FoxFeatureDestination
-import com.featuremodule.featureBImpl.ui.FoxScreen
+import com.featuremodule.foxFeatureApi.FoxFeatureDestination
+import com.featuremodule.foxFeatureImpl.ui.FoxScreen
 
 fun NavGraphBuilder.registerFoxFeature() {
     composable(route = FoxFeatureDestination.ROUTE) {
